fix(CompanyHelper): Switch Priority to Nullable Byte
Company Priority in FdsTriPpiMap is a tinyint (byte). SqlDataReader
doesn't like the explicit int cast, so converting to a byte? first, and
allowing it to implicitly cast to int?.

+++ DataRoostAPI/Access/Company/CompanyHelper.cs
@@ -486,7 +486,7 @@ namespace CCS.Fundamentals.DataRoostAPI.Access.Company {
 					using (SqlDataReader reader = cmd.ExecuteReader()) {
 						while (reader.Read()) {
 							int iconum = reader.GetInt32(0);
-							int? priority = reader.GetNullable<int>(1);
+							int? priority = reader.GetNullable<byte>(1);
 							priorityDictionary[iconum] = priority;
 						}
 					}
